requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
Brotli==1.0.9
cffi==1.15.1
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pycares==4.3.0
pycparser==2.21
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.3.2.post1
yarl==1.8.2
